refactor: add a test module for torch.functions
